Cleanup: Combine Section and BasicConfig, move to basics
