+++ src/ripple/basics/impl/Log.cpp
@@ -17,6 +17,7 @@
 */
 //==============================================================================
 
+#include <ripple/basics/Log.h>
 #include <boost/algorithm/string.hpp>
 // VFALCO TODO Use std::chrono
 #include <boost/date_time/posix_time/posix_time.hpp>
