+++ src/ripple/unity/core.cpp
@@ -25,4 +25,3 @@
 #include <ripple/core/impl/LoadMonitor.cpp>
 #include <ripple/core/impl/Job.cpp>
 #include <ripple/core/impl/JobQueue.cpp>
-#include <ripple/core/impl/Section.cpp>
